Treat users confirming preview as a request for SES registration
The idea here is to remove the need for a 'request SES registration email'
button-click - the user has confirmed they 'really' want to send to the
mailing list by clicking the link on the preview email, so we can go ahead
send them the SES verification request email without being, you know, evil.

We don't send an email if they're already registered for SES, obviously.