Allow setting the in-reply-to header on the initial patch email
Using typeahead.js for confirming the message-id - we call into the
Gmane & MARC mail archives to show a summary of the message to the user.

Added bootstrap-3-specific styling to the typeahead.js stuff. Note the
workaround for `typeahead.js-bootstrap3.less` not having quite updated to
cope with a class-name change in `typeahead.js` v0.11.1 (made the
suggestions background-transparent) - just tell typeahead.js to use
the old class name:

https://github.com/hyspace/typeahead.js-bootstrap3.less/pull/24#issuecomment-113769007

lotsa jquery, coding like it's 2010.