Appropriately set subject prefix when user sets in-reply-to
This requires parsing the subject prefix of the in-reply-to
message, and setting the new subject prefix to an incremented
version of that. E.g. '[WIP/RFC v3 0/17] Add this thing' becomes
'WIP/RFC v4' in the subject prefix field.

We use a CSS3 yellow-fade animation to show that we've changed
the subject prefix - the technique is a little more than a
decade old:

https://signalvnoise.com/archives/000558.php